Update set_background function to color each pixel according to the complex num associated with it

+++ src/mdb_set_math.cpp
@@ -1,8 +1,8 @@
 #include "mdb_set_math.h"
 
 void MdbSetMath::px_pos_to_cnum(int *px_x, int *px_y, Complex *pan, double long *zoom, Complex *cnum) {
-    cnum->setReal(pan->getReal() + *px_x * *zoom);
-    cnum->setImaginary(pan->getImaginary() - *px_y * *zoom);
+    cnum->setReal(pan->getReal() + (long double) *px_x * *zoom);
+    cnum->setImaginary(pan->getImaginary() - (long double) *px_y * *zoom);
 }
 
 void MdbSetMath::cnum_to_px_pos(Complex *cnum, Complex *pan, double long *zoom, int *px_x, int *px_y) {
